# Statistics: speed up Calendar view