Just by tweaking the SQL queries.
Also, in the hourly histogram, show at least a 1px bar when there was any reading this hour.

I always thought the slowness of showing and browsing the calendar views was caused by the many little widgets to size and draw, but no, it was really the SQL queries (and it was as slow before the recent refactoring).
On my Kobo, each of the 2 SQL queries used to take 500ms (and the UI drawing around 20ms)- and it was quite noticable.
With these tweaks, each SQL query take around 100ms.

(By adding an index `CREATE INDEX zzz ON page_stat_data(start_time)`, we'd get even better results, 50ms + 30ms - but well, it makes the DB file size grow by 30%, and it's not worth again upgrading the schema.)

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/koreader/koreader/6807)
<!-- Reviewable:end -->