@NiLuJe 2020-10-21
It's never hit a stable yet, and we've got the migration down to pat now, so I don't really have an issue with updating the schema, FWIW ;).