@poire-z 2020-10-21
OK, added the creation of that index.
A bit sad that our valuable nightly users won't get it (while our lazy or fearful :) stable users will), unless they add it manually with:
```
$ sqlite3 statistics.sqlite3
sqlite> CREATE INDEX IF NOT EXISTS page_stat_data_start_time ON page_stat_data(start_time);
sqlite> VACUUM;
```

Also added a VACUUM after migration to reclaim the space of the old deleted table page_stat. It's always a bit worrying as it may fail (but the migration too) depending on disk space or memory usage - but at least, the migration happen after a KOReader upgrade, so when the process is just launched, so probably lots of RAM available (more than after running a while).
(My DB size went from 7M to 4M with this VACUUM.)