@NiLuJe 2020-10-21
With a bit of massaging, we could arguably make `upgradeDB` safe to re-run, which would allow bumping the schema version and adding that index to the migration, too.