@NiLuJe 2020-10-21
> (Not really wanting to spend time on multiple/iterative upgrades, so I'll leave that to you if you feel like it.
> May be the safest would be to have a list of user_version tags, and just split the upgrade steps in various
> 
> ```
> function updateDBto20201010 ... end
> function updateDBto20201021 ... end
> ```
> 
> and call the functions as needed till we reach the current DB_SCHEMA_VERSION.)

Yup! Feel free to merge then, I'll look into it later tonight ;).